requirements.txt
Django==4.0.9
django-embed-video==1.4.8
django-markdownx==3.0.1
django-mathfilters==1.0.0
django-quill-editor==0.1.40
django-resized==1.0.2
django-storages==1.13.2
lxml==4.9.3
bs4==0.0.1